refactor: route groups + dual root layouts → recover full SSG
Splits the app tree so each locale gets its own root layout with a
hardcoded `<html lang>`, eliminating the `headers()` call that was
forcing every page into Dynamic SSR.

Structure
---------

  app/
    (es)/                ← route group, no URL segment
      layout.tsx         ← <html lang="es">, no headers()
      page.tsx           → /
      blog/...           → /blog/<slug>
      not-found.tsx, error.tsx, rss.xml/, llms.txt/, llms-full.txt/
    en/                  ← literal segment
      layout.tsx         ← <html lang="en">, no headers()
      page.tsx           → /en
      blog/...           → /en/blog/<slug>
      ...
    components/
      RootShell.tsx      ← shared head + body chrome (fonts, JSON-LD,
                           Analytics, SpeedInsights, LocaleSwitcher)
      blog/
        BlogClient.tsx   ← moved out of app/blog/ so both locales
                           import it cleanly without crossing groups

  proxy.ts               ← cookie redirect ONLY, no header injection;
                           matcher tightened to ["/", "/en"]

Why each piece
--------------

- Two root layouts: Next requires exactly one `<html><body>` per render,
  but route groups can have their own root layouts. (es) and en are
  siblings; each owns one branch.
- RootShell: extracts the shared chrome (fonts, person LD, Analytics)
  so neither layout duplicates 50+ lines.
- OG image files live OUTSIDE the route group at app/, app/blog/,
  app/blog/[slug]/. Inside a route group, Next 16 appends a hash
  suffix (`opengraph-image-rbbyc2`) to disambiguate, breaking the
  hardcoded URLs in our metadata. Outside the group: clean URLs.
- BlogClient moved to app/components/blog/ for clean imports from
  both locales (was app/blog/blog-client.tsx, now shared).
- proxy.ts loses its header-injection logic and shrinks to just the
  cookie redirect on bare `/` and `/en`. Matcher narrowed from
  "everything except _next/api/files" to literally those two paths,
  so /blog, /en/blog, RSS, sitemap pay zero proxy overhead.

Build output (compare to before refactor)
-----------------------------------------

Before this PR — every page was Dynamic SSR (ƒ):
  ƒ /, ƒ /blog, ƒ /blog/[slug], ƒ /en, ƒ /en/blog, ƒ /en/blog/[slug]

After this PR — every page is Static (○) or SSG (●):
  ○ /, ○ /blog, ● /blog/[slug] (4 prerendered)
  ○ /en, ○ /en/blog, ● /en/blog/[slug] (4 prerendered)
  ○ /opengraph-image, ● /blog/[slug]/opengraph-image (4 each side)

Only ƒ left are the route handlers that read getAllPosts at request
time (raw markdown, RSS, llms.txt, llms-full.txt) — those are dynamic
by design and cheap.

Verified locally
----------------

- `<html lang="es">` on /, `<html lang="en">` on /en
- /, /blog, /blog/<slug>, /opengraph-image, /blog/opengraph-image,
  /blog/<slug>/opengraph-image all serve 200
- /en mirrors all serve 200
- proxy.ts cookie redirect: bare / + cookie=en → 307 → /en;
  /en + cookie=es → 307 → /; deep links never redirect.
