+++ app/(es)/blog/layout.tsx
