+++ app/(es)/llms-full.txt/route.ts
@@ -1,7 +1,7 @@
-import { buildLlmsTxt } from "../lib/feeds";
+import { buildLlmsFullTxt } from "../../lib/feeds";
 
 export function GET() {
-  return new Response(buildLlmsTxt("es"), {
+  return new Response(buildLlmsFullTxt("es"), {
     headers: {
       "Content-Type": "text/plain; charset=utf-8",
       "Cache-Control": "public, max-age=3600, s-maxage=86400",
