+++ app/(es)/page.tsx
@@ -1,14 +1,14 @@
+import Portfolio from "../components/Portfolio";
 import { LocaleProvider } from "../lib/i18n/locale-context";
 import { getAllPosts } from "../lib/posts";
-import BlogClient from "./blog-client";
 
 const LOCALE = "es" as const;
 
-export default function BlogIndex() {
+export default function Home() {
   const posts = getAllPosts(LOCALE);
   return (
     <LocaleProvider locale={LOCALE}>
-      <BlogClient posts={posts} />
+      <Portfolio posts={posts} />
     </LocaleProvider>
   );
 }
